Fixed query support rename problems
